feat: gentler click price ladder, x2 only from 70% of the map (#110)
* feat(backend): cap the click price at x2, from 25% of the map

x3 and x4 were too harsh on a big country. With a burst of 10, a big
country still keeps 5 clicks in reserve.


* feat: price clicks in fractions of a token, x2 only from 70%

The ladder is now x1.25 from 25% of the map, x1.5 from 50% and x2 from
70%. A whole-token cost could not have a rung between x1 and x2, so the
toll, the limiter and ClickBudget.cost take a double. The proto moves
cost and next_cost to new field numbers: an old client reads zero and
says nothing about price instead of failing to decode.


---------

+++ apps/backend/internal/planet/internal/adapters/primary/http/planetv1controller/helpers_test.go
@@ -105,7 +105,7 @@ type fakeLimiter struct {
 	keys  []string
 }
 
-func (l *fakeLimiter) TakeN(key string, _ int) (bool, cpratelimit.State) {
+func (l *fakeLimiter) TakeN(key string, _ float64) (bool, cpratelimit.State) {
 	l.keys = append(l.keys, key)
 	return l.allow, l.state
 }
